apps/telegram-service/pyproject.toml
[project]
name = "telegram-service"
version = "0.1.0"
description = "Telegram gateway + admin panel for multi-tenant bot and user connections"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.36",
  "pydantic-settings>=2.6.1",
  "python-multipart>=0.0.12",
  "jinja2>=3.1.4",
  "itsdangerous>=2.2.0",
  "passlib>=1.7.4",
  "pyjwt[crypto]>=2.10.0",
  "httpx>=0.28.0",
  "google-cloud-secret-manager>=2.23.0",
  "telethon>=1.38.1"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
